Move into separate file.
